Support WadSmoosh for GZDoom's 4.6.0 changes
GZDoom 4.6.0 changed the AutoName for WadSmoosh:
doom.id.doom2.wadsmoosh -> doom.id.wadsmoosh

We support this by essentially copying the doom.id.doom2 lump filter
folder as doom.id.wadsmoosh. In order to avoid duplicating the
supershotgun sprites, however, we move these to doom.id. This is
unfortunately complicated by the fact that using the IDKFA cheat will
give the player the super shotgun if any of the sht2~ sprites exist. We
do not want this behaviour in DOOM 1 IWADs. For this reason, instead of
putting the sht2~ supershotgun sprites into the sprites folder, we put
them into a placeholder folder so that they are not automatically
recognized and imported as sprites. We can then use a textures lump to
re-import them as sprites only within the doom.id.doom2 and
doom.id.wadsmoosh lump filters.

See @madame-rachelle for this method, as used in WidePix:
nashmuhandes/WidePix@89dac47
https://forum.zdoom.org/viewtopic.php?p=1190878#p1190878